requirements.txt
scrapy>=2.5.0
selenium>=4.0.0
psycopg2-binary>=2.9.0
pandas>=1.5.0
schedule>=1.2.0
beautifulsoup4>=4.11.0
requests>=2.28.0
python-dateutil>=2.8.0
webdriver-manager>=3.8.0
lxml>=4.9.0
itemadapter>=0.7.0
twisted>=22.10.0
pytz>=2022.7
sqlalchemy>=1.4.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0